+++ streampipes-extensions/streampipes-connectors-mqtt/src/main/java/org/apache/streampipes/extensions/connectors/mqtt/security/SecurityUtils.java
@@ -1,0 +1,157 @@
+package org.apache.streampipes.extensions.connectors.mqtt.security;
+
+import javax.net.ssl.*;
+
+import org.apache.streampipes.commons.environment.Environments;
+
+import java.io.*;
+import java.net.Socket;
+import java.security.*;
+import java.security.cert.*;
+import java.security.spec.PKCS8EncodedKeySpec;
+import java.util.Base64;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+public class SecurityUtils {
+
+    public static TrustManager[] acceptAllCerts() {
+        return new TrustManager[] {
+                new X509TrustManager() {
+                    public X509Certificate[] getAcceptedIssuers() {
+                        return null;
+                    }
+
+                    public void checkClientTrusted(X509Certificate[] certs, String authType) {
+                    }
+
+                    public void checkServerTrusted(X509Certificate[] certs, String authType) {
+                    }
+                }
+        };
+    }
+
+    public static X509Certificate getServerCertificate(String host, int port)
+            throws NoSuchAlgorithmException, IOException, CertificateException {
+
+        SSLSocketFactory factory = (SSLSocketFactory) SSLSocketFactory.getDefault();
+        try (Socket socket = factory.createSocket(host, port)) {
+
+            ((SSLSocket) socket).startHandshake();
+
+            SSLSession session = ((SSLSocket) socket).getSession();
+            var certChain = session.getPeerCertificates();
+
+            return (X509Certificate) certChain[0];
+        }
+    }
+
+    public static KeyStore loadServerKeyStore() throws FileNotFoundException, KeyStoreException, IOException,
+            NoSuchAlgorithmException, CertificateException {
+
+             var env = Environments.getEnvironment();
+        String keystoreFilename = env.getKeystoreFilename().getValueOrDefault();
+        String keystoreType = env.getKeystoreType().getValueOrDefault();
+        String keystorePassword = env.getKeystorePassword().getValueOrDefault();
+
+        try (FileInputStream keystoreFile = new FileInputStream(keystoreFilename)) {
+            KeyStore keystore = KeyStore.getInstance(keystoreType);
+            keystore.load(keystoreFile, keystorePassword.toCharArray());
+            return keystore;
+        } catch (IOException | NoSuchAlgorithmException | CertificateException e) {
+            throw e;
+        }
+    }
+
+    public static KeyManager[] loadClientKeyManagers(String certPem, String keyPem) throws Exception {
+        X509Certificate certificate = parseCertificateFromPem(certPem);
+        PrivateKey privateKey = parsePrivateKeyFromPem(keyPem);
+
+        String password = ""; // no password for in-memory keystore
+        KeyStore keyStore = KeyStore.getInstance("PKCS12");
+        keyStore.load(null, null);
+        keyStore.setKeyEntry("client", privateKey, password.toCharArray(), new java.security.cert.Certificate[]{certificate});
+
+        KeyManagerFactory kmf = KeyManagerFactory.getInstance(KeyManagerFactory.getDefaultAlgorithm());
+        kmf.init(keyStore, password.toCharArray());
+        return kmf.getKeyManagers();
+    }
+
+    public static X509Certificate parseCertificateFromPem(String pem) throws Exception {
+        String normalized = pem
+                .replace("-----BEGIN CERTIFICATE-----", "")
+                .replace("-----END CERTIFICATE-----", "")
+                .replaceAll("\\s+", "");
+        byte[] decoded = Base64.getDecoder().decode(normalized);
+        CertificateFactory cf = CertificateFactory.getInstance("X.509");
+        return (X509Certificate) cf.generateCertificate(new ByteArrayInputStream(decoded));
+    }
+
+    public static PrivateKey parsePrivateKeyFromPem(String pem) throws Exception {
+        pem = pem.replace("\\n", "\n")
+                 .replace("\\r", "")
+                 .replace("\r", "")
+                 .trim();
+
+        if (pem.contains("-----BEGIN RSA PRIVATE KEY-----")) {
+            return parsePkcs1PrivateKey(pem);
+        } else {
+            throw new IllegalArgumentException("Unsupported key format: missing BEGIN/END markers");
+        }
+    }
+
+    public static PrivateKey parsePkcs1PrivateKey(String pem) throws Exception {
+        Pattern p = Pattern.compile(
+            "-----BEGIN RSA PRIVATE KEY-----([A-Za-z0-9+/=\\s]+)-----END RSA PRIVATE KEY-----"
+        );
+        Matcher m = p.matcher(pem.replaceAll("\\s+", ""));
+        if (!m.find()) {
+            pem = pem.replaceAll("-----BEGIN RSA PRIVATE KEY-----", "")
+                     .replaceAll("-----END RSA PRIVATE KEY-----", "")
+                     .replaceAll("[\\s\\r\\n]", "")
+                     .trim();
+        } else {
+            pem = m.group(1);
+        }
+
+        byte[] pkcs1Bytes = Base64.getMimeDecoder().decode(pem);
+        byte[] pkcs8Bytes = convertPkcs1ToPkcs8(pkcs1Bytes);
+
+        PKCS8EncodedKeySpec keySpec = new PKCS8EncodedKeySpec(pkcs8Bytes);
+        KeyFactory keyFactory = KeyFactory.getInstance("RSA");
+        return keyFactory.generatePrivate(keySpec);
+    }
+
+    private static byte[] convertPkcs1ToPkcs8(byte[] pkcs1Bytes) throws IOException {
+        final byte[] pkcs8Header = new byte[] {
+            0x30, (byte)0x82, 
+            0, 0, 
+            0x02, 0x01, 0x00,
+            0x30, 0x0d,
+            0x06, 0x09,
+            0x2a, (byte)0x86, 0x48, (byte)0x86, (byte)0xf7, 0x0d, 0x01, 0x01, 0x01,
+            0x05, 0x00,
+            0x04, (byte)0x82,
+            0, 0
+        };
+
+        int pkcs1Length = pkcs1Bytes.length;
+        int totalLength = pkcs8Header.length + pkcs1Length;
+
+        pkcs8Header[2] = (byte)((totalLength - 4) >> 8);
+        pkcs8Header[3] = (byte)(totalLength - 4);
+        pkcs8Header[pkcs8Header.length - 2] = (byte)(pkcs1Length >> 8);
+        pkcs8Header[pkcs8Header.length - 1] = (byte)(pkcs1Length);
+
+        ByteArrayOutputStream out = new ByteArrayOutputStream();
+        out.write(pkcs8Header);
+        out.write(pkcs1Bytes);
+        return out.toByteArray();
+    }
+
+    public static TrustManagerFactory createTrustManagerFactory(KeyStore keystore) throws Exception {
+        TrustManagerFactory trustManagerFactory = TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
+        trustManagerFactory.init(keystore);
+        return trustManagerFactory;
+    }
+}
